backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
stripe==10.12.0
redis==5.1.1
aiofiles==24.1.0
openpyxl==3.1.5
python-dotenv==1.0.1
httpx==0.27.2
slowapi==0.1.9
